+++ src/_incydr_cli/cmds/user_risk_profiles.py
@@ -1,20 +1,12 @@
-from functools import lru_cache
-from pathlib import Path
 from typing import Optional
 
 import click
-from pydantic import Field
-from requests import HTTPError
 from rich.panel import Panel
-from rich.progress import track
 
 from _incydr_cli import console
 from _incydr_cli import logging_options
 from _incydr_cli import render
-from _incydr_cli.cmds.models import UserCSV
-from _incydr_cli.cmds.models import UserJSON
 from _incydr_cli.cmds.options.output_options import columns_option
-from _incydr_cli.cmds.options.output_options import input_format_option
 from _incydr_cli.cmds.options.output_options import single_format_option
 from _incydr_cli.cmds.options.output_options import SingleFormat
 from _incydr_cli.cmds.options.output_options import table_format_option
@@ -22,7 +14,6 @@ from _incydr_cli.cmds.options.output_options import TableFormat
 from _incydr_cli.cmds.options.profile_filter_options import profile_filter_options
 from _incydr_cli.cmds.options.utils import user_lookup_callback
 from _incydr_cli.cmds.users import users
-from _incydr_cli.cmds.utils import user_lookup
 from _incydr_cli.core import incompatible_with
 from _incydr_cli.core import IncydrCommand
 from _incydr_cli.core import IncydrGroup
@@ -30,14 +21,6 @@ from _incydr_sdk.core.client import Client
 from _incydr_sdk.exceptions import DateParseError
 from _incydr_sdk.user_risk_profiles.models import UserRiskProfile
 from _incydr_sdk.utils import model_as_card
-
-
-class UpdateCloudAliasesCSV(UserCSV):
-    cloud_alias: str = Field(csv_aliases=["cloudAlias", "cloud_alias", "alias"])
-
-
-class UpdateCloudAliasesJSON(UserJSON):
-    cloud_alias: str = Field(alias="cloudAlias")
 
 
 @users.group(cls=IncydrGroup)
@@ -218,140 +201,3 @@ def update(
             console.print(updated_profile.json(), highlight=False)
     except DateParseError as err:
         raise err
-
-
-@risk_profiles.command(cls=IncydrCommand)
-@click.argument("user", callback=user_lookup_callback)
-@click.argument("cloud-alias")
-@logging_options
-def add_cloud_alias(
-    user: str,
-    cloud_alias: str,
-):
-    """
-    Add a cloud alias to a user risk profile.
-
-    Accepts a user ID or a username.  Performs an additional lookup if a username is passed.
-
-    A cloud alias is the username an employee uses to access cloud services such as Google Drive or Box.
-    Adding a cloud alias allows Incydr to link a user's cloud activity with their Code42 username.
-    Each user has a default cloud alias of their Code42 username. You can add one additional alias.
-    """
-    client = Client()
-    client.user_risk_profiles.v1.add_cloud_alias(user, cloud_alias)
-    console.print(f"Cloud alias successfully added to user '{user}")
-
-
-@risk_profiles.command(cls=IncydrCommand)
-@click.argument("user", callback=user_lookup_callback)
-@click.argument("cloud-alias")
-@logging_options
-def remove_cloud_alias(
-    user: str,
-    cloud_alias: str,
-):
-    """
-    Remove a cloud alias from a user risk profile.
-
-    Accepts a user ID or a username.  Performs an additional lookup if a username is passed.
-    """
-    client = Client()
-    client.user_risk_profiles.v1.delete_cloud_alias(user, cloud_alias)
-    console.print(f"Cloud alias successfully removed from user '{user}")
-
-
-@risk_profiles.command(cls=IncydrCommand)
-@click.argument("file", type=click.File())
-@input_format_option
-@logging_options
-def bulk_add_cloud_aliases(file: Path, format_: str):
-
-    """
-    Bulk add cloud aliases to user risk profiles.
-
-    Takes a single arg `FILE` which specifies the path to the file (use "-" to read from stdin).
-
-    File format can either be CSV or [JSON Lines format](https://jsonlines.org) (Default is CSV).
-
-    Required CSV columns:
-
-     * `user` - User ID or username of the user risk profile. Performs an additional lookup if username is passed.
-     * `cloud_alias` - The cloud alias to remove from the profile.
-    """
-    client = Client()
-
-    @lru_cache()
-    def resolve_username(user):
-        if user is None:
-            return
-        elif "@" in user:
-            return user_lookup(client, user)
-        else:  # assume user_id
-            return user
-
-    if format_ == "csv":
-        models = UpdateCloudAliasesCSV.parse_csv(file)
-    else:
-        models = UpdateCloudAliasesJSON.parse_json_lines(file)
-
-    try:
-        for row in track(
-            models,
-            description="Adding cloud aliases...",
-            transient=True,
-        ):
-            client.user_risk_profiles.v1.add_cloud_alias(
-                resolve_username(row.user), row.cloud_alias
-            )
-    except ValueError as err:
-        console.print(f"[red]Error:[/red] {err}")
-    except HTTPError as err:
-        console.print(f"[red]Error:[/red] {err.response.text}")
-
-
-@risk_profiles.command(cls=IncydrCommand)
-@click.argument("file", type=click.File())
-@input_format_option
-@logging_options
-def bulk_remove_cloud_aliases(file: Path, format_: str):
-    """
-    Bulk remove cloud aliases from user risk profiles.
-
-    Takes a single arg `FILE` which specifies the path to the file (use "-" to read from stdin).
-
-    File format can either be CSV or [JSON Lines format](https://jsonlines.org) (Default is CSV).
-
-    Required CSV columns:
-
-     * `user` - User ID or username of the user risk profile. Performs an additional lookup if username is passed.
-     * `cloud_alias` - The cloud alias to remove from the profile.
-    """
-    client = Client()
-
-    @lru_cache()
-    def resolve_username(user):
-        if user is None:
-            return
-        elif "@" in user:
-            return user_lookup(client, user)
-        else:  # assume user_id
-            return user
-
-    if format_ == "csv":
-        models = UpdateCloudAliasesCSV.parse_csv(file)
-    else:
-        models = UpdateCloudAliasesJSON.parse_json_lines(file)
-
-    try:
-        for row in track(
-            models,
-            description="Removing cloud aliases...",
-            transient=True,
-        ):
-            client.user_risk_profiles.v1.delete_cloud_alias(
-                resolve_username(row.user), row.cloud_alias
-            )
-    except ValueError as err:
-        console.print(f"[red]Error:[/red] {err}")
-    except HTTPError as err:
-        console.print(f"[red]Error:[/red] {err.response.text}")